fix(shared-ui): move formatCpfProgressive de shared-data para eliminar import cross-lib
shared-ui:build:production falhava com TypeError em getReferencedFileLocation
porque ng-packagr não lida bem com import cross-lib via TS path mapping
(@uniplus/shared-data resolvendo para source de outra lib).

formatCpfProgressive é formatador puro de UI (sem semântica de domínio),
único consumidor era cpf-input.ts em shared-ui. Movido para
libs/shared-ui/src/lib/utils/cpf-format.util.ts com import relativo.
Spec dedicada migrada junto.

Refs #124
